Update Product Switcher product images and customize current application name (#3)
This commit adds a few new product icon assets for Dispatch, Airstrip. This so we can distinguish it from just the Firmhouse platform (previously known as GoMonthly). That way we can position Airstrip and Dispatch more clearly as separate products that Firmhouse also offers, but they are not core of our platform.

Also added the option to pass in a `product_name` property argument. This is then the text that shows up in the product switcher as "current product"

Co-authored-by: Michiel Sikkes <michiel@firmhouse.com>